refactor controller to be own subsystem - to support possible rumbling for duration - add feed shot facing amp corner todo - add amp facing command - remove drive to speaker preset
