+++ src/main/java/frc/robot/Robot.java
@@ -5,10 +5,8 @@
 package frc.robot;
 
 import edu.wpi.first.wpilibj.TimedRobot;
-import edu.wpi.first.wpilibj.GenericHID.RumbleType;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.robot.subsystems.Feeder;
 import io.github.oblarg.oblog.Logger;
 
 /**
@@ -86,11 +84,7 @@ public class Robot extends TimedRobot {
 
   /** This function is called periodically during operator control. */
   @Override
-  public void teleopPeriodic() {
-    double rumbleValue = Feeder.getInstance().gamePieceDetected() ? 1.0 : 0.0;
-    m_robotContainer.m_driverController.getHID().setRumble(RumbleType.kBothRumble, rumbleValue);
-    m_robotContainer.m_operatorController.getHID().setRumble(RumbleType.kBothRumble, rumbleValue);
-  }
+  public void teleopPeriodic() {}
 
   @Override
   public void testInit() {
